AuthContext, Login-method & NavBar
Added a useContext-hook for storing authentication-data.

Created a page for login-prompt that acts as the landing page for the employee portal. The page has a login form that sends an API-call to the server to authenticate the user.

If authenticated the employee gets routed to MyPages which currently only has a simple navbar and displays the id of the current user as a confirmation that the login was successful.

The logout button simply clears the data in the authContext and navigates the user back to the login-view.

Changed the port to 3000 since the other front-end project is already running on the standard port 5173.