JobsTable
Tried implementing React Query but decided that this probably won't be beneficial unless the database becomes much larger. Until then it's probably better to just fetch all the jobs when the component mounts and then render the table based on what kinds of jobs the admin wants to administrate.

Created a Checkbox-component and a StatusFilter-component for separating the logic for filtering the table of jobs.

The AdminPages will by default show the jobs that has the status of OPEN and NOT_APPROVED - since these are the jobs that an admin probably first want to deal with when signing in to the portal.

The Assign employee(s)-button is still just a button. The logic for that one will be implemented in a later user story.