AdminPages - StatusFilter (WIP)
Moved the filtering checkboxes in a dropdown-button instead. The dropdown contains 3 columns with different filtering criteria (status, type and employee). On smaller screens it's presented in just one long column instead.

Added a "Select all"-button and a "Clear filter"-button for easier filtering.

Still need to implement the logic for actually filtering the jobs by job type and employee. Right now it only has the capacity of filtering based on job status.