Assign employee(s) to a job (WIP)
Added the logic for assigning an employee to a job. The "Assign employee(s)" now calls on a modal that fetches all the employees that don't have any assignment on this day (if the job is set to NOT_APPROVED the assigned cleaners will be shown as available as well).

The employees are presented in the SelectEmployees-component. It's possible to choose more than one employee for the job if needed.

There's still some bugs left to fix (like the button not displaying the modal on the first click, and not disappearing when the job is assigned), but the actual API call works fine. The job is assigned to the requested employees, which triggers a re-fetch of the list of jobs to render the new status.