BUGFIX: Assign employee(s) to a job
Created a new modal using react-bootstrap instead since there seems to be some issues with how BS and React handles the JS-code for showing or hiding the modal.

The modal now shows up as expected on the first click, and it is possible to assign employees.

The modal will now close when an employee is assigned to the job.

Added some placeholder data for the table while fetching data for the jobs as well as a spinner for the assign-button while waiting for the server to respond on the PUT-request.