REFACTOR: Clean up on AdminPages
Removed some duplicates from the AdminPages. The NavBar for example was duplicated in an earlier commit.

Refactored the different headings and containers for a unified look.

These components should probably be rendered on different views later on since there's going to be a lot of data fetching and the AdminPages-view will be overcrowded and harder to read.