CREATED: FormRegisterEmployee
Added two more views for registration of new employees - one for creating a new admin and the other for creating new cleaners. I may move these to modals instead - I'm not sure what's the best solution at this time.

Made a small change to the NavBar as well by making the "logo" a link back to the dashboard. This should also be replaced with a logo, hence the name-change.

The API-call for creating new employees work as expected.