BUGFIX: FormUpdateCustomer
Fixed a bug in the API-call that had a couple of misspelling that causes some of the fields to not be updated.

I changed the current values to be set as defaultValues instead of placeholder. This way the fields display and hold the actual data - but does work the same way as before.

Also added a cancel-button to allow the admin to go back to the dashboard if the edit-button was clicked on by mistake.