REFACTOR: FormUpdateCustomer
Made the API-call accept optional values instead and moved the logic for checking whether a field has been updated or not to the form-component instead.

Now the API-call will only send an PUT-request with the updated fields as it did before moving the current value from placeholder to defaultValues, but the fields will still be prefilled with the current values.