CREATED: EmployeeTab
Created a tab with employee-data. The tab currently includes a table for displaying all admins (re-used the component for displaying customers).

Need to implement the logic for updating, with a new form component and the logic for deleting an admin.

There's also room for adding a table for administration of cleaners.