CREATED: CleanersTable
Created the table for listing cleaners on the EmployeeTab.

Implemented the methods for deleting a cleaner with an alert message if the deletion is unsuccessful due to jobs in progress.

Added a form for updating the cleaner's data. The form is re-used from the customer-form and is generic - with the intention of re-use again by the AdminsTable. Since the data for the admins and cleaners are the same except for the role you only have to provide the role with the values for the headings to change accordingly.