REFACTOR: FormField, FormRegister... & FormUpdateEmployee
Updated the FormField to use light background on the input fields for better visibility on the dark background.

Put all input-fields on a separate line. There's no need to compress the form since there are already few fields.

Made some smaller design changes to the headings color and font-weight etc.