CREATED: API-call for InvoiceTable
Added the API-call for handling the mark-as-paid request as well as a request to delete an invoice.

Added a new column for handling the delete and move the button for updating the status to the status-column. The status column now renders the status if the job has been paid, otherwise the button for updating the status will show up in this column.

Had to tweak the modal a bit to get it to work as expected and to handle both "actions". Refactored the modal to work with react-bootstraps component.