CREATED: JWT implementation
Added an access-token to all API-calls. Store it in the session-storage for now. Probably want to store this in a cookie with http-only flag later on, but has to make some changes to the server for that to work.

The role variable was removed from the context and is now stored in the session-storage as well.