REFACTOR: MyPages
The API-calls in each useEffect-hook now waits for the user to switch to the tab to run - so the data won't be fetched if the tab is not active.

This also means that the useEffect will make sure that the data shown to the user always is updated since it will fetch new data every time the tab is active.