REFACTOR: EditEmployeeDataView
Made some changes to the styling of this view since there were some extra divs and containers that caused some issues with unnecessary padding and margins.

Removed the button to cancel in the update-employee form since it is not needed when presented together with the form for updating password. It makes more sense to have only one button to go back to the dashboard.

Translated the password-form to English since the rest of the employee-project is written in English.