REFACTOR: Job tables
Made the custom ID partially hidden because all the data in the table didn't fit the container else now that we also renders the date and time of each job.

The user may choose to click on the customer ID to expand the full ID.

Added the ability to show and hide messages as well in the cleaners table.

Moved the settings-button to the navbar as well, and removed  the WIP-tab from the admin pages that is no longer in use.